usb: xhci-msm-hsic: Add support for pinctrl framework
As the newer targets will be using pinctrl framework for gpio
configuration, our drivers need to be compatible with both set of targets
the one which do not use pinctrl and one which do. Adding support for both
these cases. xhci-hsic driver can configure data, strobe and host-ready
gpio's.

Change-Id: Ife1b04ba14013cc62896831e0be50a84b71ab6f0
Signed-off-by: Tarun Gupta <tarung@codeaurora.org>

Conflicts:
	drivers/usb/host/xhci-msm-hsic.c