usb: dwc3: Add support to disable hsphy auto suspend feature
Currently hsphy auto suspend is enabled by default. There are few
issues using this feature as it is required to disable this
functionality before sending different endpoint and generic commands
to core and re-enable after the same. Hence this change adds support
to disable hsphy auto suspend feature if it is required.

CRs-Fixed: 625668
Change-Id: Iec585aefe83fb5710152dd6e4627fcba767cd262
Signed-off-by: Mayank Rana <mrana@codeaurora.org>

Conflicts:
	drivers/usb/dwc3/core.h