msm: ipa: introduce new QMI msg for IPA config
One of the common issues that we face when debugging CAT-6
throughput issues for tethered calls is because the USB is
not connected in super speed mode (USB 3.0). So, we would
like to capture this info in the MDM crashdumps because
it will ease our analysis of the issue. Apps will send
this QMI message to modem for 9x35 LE & 8994 to inform
the modem about the USB core speed.

Change-Id: Ia3cf29fae6ab13ad49384219d7bdd9fbc03c6219
Signed-off-by: Skylar Chang <chiaweic@codeaurora.org>

Conflicts:
	include/uapi/linux/ipa_qmi_service_v01.h